requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
boto3>=1.34
stripe>=8.0
gunicorn>=21.2
django-filter>=23.5
argon2-cffi>=23.1
